researcher/pyproject.toml
[project]
name = "researcher-agent"
version = "0.1.0"
description = "Researcher microservice agent"
requires-python = ">=3.10,<3.14"
dependencies = [
    "google-adk==1.18.0",
    "fastapi==0.128.0",
    "uvicorn==0.34.0",
    "google-cloud-logging==3.12.0",
    "opentelemetry-exporter-gcp-trace==1.9.0",
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
